[juju] Add juju cluster profile to support juju managed clouds.
Add a new cluster profile for `sos collect`. This profile is intended to
be used on juju managed environments which assumes that `juju` is
installed on the machine where `sos collect` is called, and that the
juju user has superuser privilege to the current controller.

By default, the sos reports will be collected from all the applications
in the current model. If necessary, one can filter the nodes by models /
applications / units / machines with cluster options. For example `-c
"juju.models=sos" -c "juju.apps=a,b,c"`

Closes: #2668

Signed-off-by: Chi Wai Chan <chiwai.chan@canonical.com>