[juju] Add support for juju ssh transport.
This transport is expected to be used in juju managed environment, and
the user should have the necessary credential for accessing the
controller.  When using this transport, the --nodes option will be
expected to be a comma separated machine IDs or unit names, **not** IP
addr, since `juju ssh` identifies the ssh target by machine ID or unit
name. For example, `sos collect --nodes 0,1,2 --transport juju --batch`
and `sos collect --nodes ubuntu/0,ubuntu/1 --transport juju`.

Closes: #2668

Signed-off-by: Chi Wai Chan <chiwai.chan@canonical.com>